example-voting-app/load-testing/requirements.txt
Flask==0.10.1
gevent==1.1.1
greenlet==0.4.9
itsdangerous==0.24
Jinja2==2.8
locustio==0.7.3
MarkupSafe==0.23
msgpack-python==0.4.7
requests==2.9.1
Werkzeug==0.11.8
